#44 Moved itde fixtures to __init__.py
